+++ packages/ui/multiInputField/src/MultiInputField.tsx
@@ -22,7 +22,7 @@ import useMergedRef from '@react-hook/merged-ref';
 import { Root } from '@radix-ui/react-label';
 import { Description, Label, TextError } from '@pikas-ui/text';
 import { Tooltip, TooltipCSS } from '@pikas-ui/tooltip';
-import fontColorContrast from 'font-color-contrast';
+import { Color } from '@pikas-utils/color';
 import { MultiInputFieldBadge } from './badge/index.js';
 
 const Container = styled('div', {
@@ -300,10 +300,9 @@ const getColor = ({
   } else if (colorName) {
     return `$${colorName}`;
   } else if (theme) {
-    return fontColorContrast(
-      theme.colors[backgroundColorName ?? 'WHITE'].value,
-      0.7
-    );
+    return new Color(
+      theme.colors[backgroundColorName ?? 'WHITE'].value
+    ).getContrast();
   } else {
     return '';
   }