+++ packages/ui/alert/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pikas-ui/alert
 
+## 3.4.9
+
+### Patch Changes
+
+- Upgrade dependencies
+- Fix ESM
+- Updated dependencies
+- Updated dependencies
+  - @pikas-ui/icons@3.3.9
+  - @pikas-ui/styles@3.5.9
+
 ## 3.4.8
 
 ### Patch Changes