+++ packages/ui/avatar/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pikas-ui/avatar
 
+## 2.3.9
+
+### Patch Changes
+
+- Upgrade dependencies
+- Fix ESM
+- Updated dependencies
+- Updated dependencies
+  - @pikas-ui/skeleton@2.2.9
+  - @pikas-ui/styles@3.5.9
+
 ## 2.3.8
 
 ### Patch Changes