+++ packages/ui/badge/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pikas-ui/badge
 
+## 1.3.10
+
+### Patch Changes
+
+- Upgrade dependencies
+- Fix ESM
+- Updated dependencies
+- Updated dependencies
+  - @pikas-ui/icons@3.3.9
+  - @pikas-ui/styles@3.5.9
+
 ## 1.3.9
 
 ### Patch Changes