+++ packages/ui/dialog/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @pikas-ui/dialog
 
+## 2.4.13
+
+### Patch Changes
+
+- Upgrade dependencies
+- Fix ESM
+- Updated dependencies
+- Updated dependencies
+  - @pikas-ui/button@3.3.9
+  - @pikas-ui/icons@3.3.9
+  - @pikas-ui/loader@3.3.9
+  - @pikas-ui/slider@3.3.10
+  - @pikas-ui/styles@3.5.9
+  - @pikas-ui/title@1.2.10
+
 ## 2.4.12
 
 ### Patch Changes