+++ packages/ui/dropdown-menu/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @pikas-ui/dropdown-menu
 
+## 3.4.13
+
+### Patch Changes
+
+- Upgrade dependencies
+- Fix ESM
+- Updated dependencies
+- Updated dependencies
+  - @pikas-ui/icons@3.3.9
+  - @pikas-ui/loader@3.3.9
+  - @pikas-ui/menu@3.3.13
+  - @pikas-ui/styles@3.5.9
+
 ## 3.4.12
 
 ### Patch Changes