+++ packages/ui/multiInputField/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @pikas-ui/multi-input-field
 
+## 2.3.11
+
+### Patch Changes
+
+- Upgrade dependencies
+- Fix ESM
+- Updated dependencies
+- Updated dependencies
+  - @pikas-ui/badge@1.3.10
+  - @pikas-ui/icons@3.3.9
+  - @pikas-ui/styles@3.5.9
+  - @pikas-ui/text@4.2.10
+  - @pikas-ui/tooltip@3.3.9
+
 ## 2.3.10
 
 ### Patch Changes