+++ packages/ui/text/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pikas-ui/text
 
+## 4.2.10
+
+### Patch Changes
+
+- Upgrade dependencies
+- Fix ESM
+- Updated dependencies
+- Updated dependencies
+  - @pikas-ui/styles@3.5.9
+
 ## 4.2.9
 
 ### Patch Changes